backend/requirements.txt
fastapi
uvicorn
endee
sentence-transformers
python-multipart
pypdf
python-dotenv
requests
google-generativeai
